+++ DapperExtensionsDNXCore.Test/IntegrationTests/Sqlite/CrudFixture.cs
@@ -17,7 +17,7 @@ namespace DapperExtensions.Test.IntegrationTests.Sqlite
             public void AddsEntityToDatabase_ReturnsKey()
             {
                 Person p = new Person { Active = true, FirstName = "Foo", LastName = "Bar", DateCreated = DateTime.UtcNow };
-                int id = Db.Insert(p);
+                int id = (int)Db.Insert(p);
                 Assert.AreEqual(1, id);
                 Assert.AreEqual(1, p.Id);
             }
@@ -69,7 +69,7 @@ namespace DapperExtensions.Test.IntegrationTests.Sqlite
                     LastName = "Bar",
                     DateCreated = DateTime.UtcNow
                 };
-                int id = Db.Insert(p1);
+                int id = (int)Db.Insert(p1);
 
                 Person p2 = Db.Get<Person>(id);
                 Assert.AreEqual(id, p2.Id);
@@ -103,7 +103,7 @@ namespace DapperExtensions.Test.IntegrationTests.Sqlite
                     LastName = "Bar",
                     DateCreated = DateTime.UtcNow
                 };
-                int id = Db.Insert(p1);
+                int id = (int)Db.Insert(p1);
 
                 Person p2 = Db.Get<Person>(id);
                 Db.Delete(p2);
@@ -176,7 +176,7 @@ namespace DapperExtensions.Test.IntegrationTests.Sqlite
                     LastName = "Bar",
                     DateCreated = DateTime.UtcNow
                 };
-                int id = Db.Insert(p1);
+                int id = (int)Db.Insert(p1);
 
                 var p2 = Db.Get<Person>(id);
                 p2.FirstName = "Baz";
